refactor: 패키지명 수정
기존 model이던 패키지명을 domain으로 수정합니다.
