+++ src/main/java/attendance/domain/AttendanceBook.java
@@ -1,4 +1,4 @@
-package attendance.model;
+package attendance.domain;
 
 import java.time.LocalDate;
 import java.util.ArrayList;
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.Map;
 
 public class AttendanceBook {
-    
+
     private final Map<String, List<AttendanceTime>> attendances = new HashMap<>();
 
     public void add(final String name, final AttendanceTime attendanceTime) {
