+++ src/test/java/AttendanceBookTest.java
@@ -1,8 +1,8 @@
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
-import attendance.model.AttendanceBook;
-import attendance.model.AttendanceStatus;
-import attendance.model.AttendanceTime;
+import attendance.domain.AttendanceBook;
+import attendance.domain.AttendanceStatus;
+import attendance.domain.AttendanceTime;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.stream.Stream;
